requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
httpx==0.25.1
python-dotenv==1.0.0

